+++ htdocs/js/05_athleten.js
@@ -531,7 +531,8 @@ function _athSetFullPage(id, slug) {
   closeModal(); syncHash(); buildNav(); renderPage();
 }
 function _athCopyLink(slug) {
-  var url = location.origin + location.pathname + '#athlet/' + slug;
+  // Pfad-basierte URL (/athlet/slug) – Messenger/Crawler lesen OG-Tags aus athlet.php
+  var url = location.origin + '/athlet/' + slug;
   if (navigator.clipboard && navigator.clipboard.writeText) {
     navigator.clipboard.writeText(url).then(function(){ notify('Link kopiert!','ok'); });
   } else {